\name{pd.filtered.tetexpr}
\Rdversion{1.1}
\docType{package}
\alias{pd.filtered.tetexpr}
\title{
  Annotation package for pd.filtered.tetexpr.
}
\description{
  Annotation package for pd.filtered.tetexpr built with pdInfoBuilder.
}
\details{
  This package is to be used in conjunction with the oligo package.
}
\keyword{package}
